+++ lib/ua_inspector/short_code_map/desktop_families.ex
@@ -1,0 +1,14 @@
+defmodule UAInspector.ShortCodeMap.DesktopFamilies do
+  @moduledoc """
+  Desktop Families Map.
+  """
+
+  use UAInspector.ShortCodeMap,
+    ets_prefix: :ua_inspector_scm_desktop_families,
+    file_local: "short_codes.desktop_families.yml",
+    file_remote: "DeviceDetector.php",
+    var_name: "desktopOsArray",
+    var_type: :list
+
+  def to_ets(item), do: {item}
+end