+++ test/fixtures/DeviceDetector.php
@@ -1,0 +1,6 @@
+<?php
+
+class DeviceDetector
+{
+    protected static $desktopOsArray = array('Windows');
+}