+++ CHANGELOG.md
@@ -9,6 +9,8 @@
       of calling the database state server
     - If you need to check if all databases are loaded (i.e. "no longer empty")
       you can use `UAInspector.ready?/0`
+    - Reloading will now issue a warning if no database path is configured
+      while resuming operation with an empty database
 
 - Soft deprecations (no warnings)
     - The individual download tasks for databases and short code maps have